chg: dev: filter genres and delete unuse fields
.gitignore: allow filtered json file
src/assets/filter_json.py: filter out the game which doen't have any one of the genres in the
top10 list.
If a game has multiple genres in the top10 list, look up the tags and choose the one with the
highest votes. If tags field doesn't exist or tags do not include top10 genres, drop that
game.
If a game doesn't have genres in the top10, simply drop that game.